⬆️ net-misc/httpie: bump to 3.2.3
Note that there is no <dev-python/requests-3.32.3 for Python 3
in the portage tree at the moment, so requests' version can't
be effectively pinned like it was pinned in the PyPI package.
It leads to errors when processing HTTPS links.
Upcoming release of Requests should fix it.
Issue for tracking:
  - https://github.com/psf/requests/issues/6730

Release:
  - https://github.com/httpie/cli/releases/tag/3.2.3

Changes in comparison to 3.2.2:
  - add Python 3.13 support
  - add the repository url to the HOMEPAGE